Network created
Created the class PlaneBrain which is the network which controls a plane. It has 4 arrays for the 4 different layers.  Also made edits to the BrainNode class: added enum for their type(which layer are they in). Also changed Arraylists to just arrays because I will not be adding to them so it is not require for them to be Arraylists.